cilium: dockerfiles for llvm, bpftool and iproute2
Split-off from the main repo to have them located under cilium/packaging.
Starting from ubuntu:20.04 as base image since we're migrating there
anyway.

Also, remove the ARCH setting so it can be done properly for each case.
For example, seen the following on quay.io for bpftool:

  [...]
  Makefile:673: arch/amd64/Makefile: No such file or directory
  make: *** No rule to make target 'arch/amd64/Makefile'. Stop.
  [...]

This seems to collide with kernel ARCH setting in the makefile. Remove
it for now to avoid clashes. This needs to be revisited.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>