Ensure images are always re-built in local mode
There is no need to check for local presence, as if it was built on
the same host recently, the build cache will be used and it should
be very quick.

The build was forced in PR mode and otherwise on a local branch,
due to `-dev` suffix, but it broke on master.

Signed-off-by: Ilya Dmitrichenko <errordeveloper@gmail.com>