Fixing mysterious 400 errors with quay.io
There appears to be a weird side-effect of using `FROM scratch`
with Quay, here is an illustration:

https://gist.github.com/errordeveloper/7a078c53d4694f0cad9a727255fb23a6

It turns out that adding an empty file fixes this. This had been
reported to Quay team, and the workaround can hopefully get reverted
sometime. It also appears to be specific to BuildKit.

This commits stops short of pushing compilers and llvm images to
Quay, as there are bigger and it's best to ensure workaround works
for all of the other images first that are quick to re-build.

Signed-off-by: Ilya Dmitrichenko <errordeveloper@gmail.com>