checkpatch: prepare for new image, add checkpatch.pl from Linux
In preparation for a new image in the repository, create a new directory
image/checkpatch, and populate it with the checkpatch.pl script from the
Linux repository. The accompanying license and spelling files are also
included. All files are taken from kernel commit f436a58e2619
("checkpatch: fix CONST_STRUCT when const_structs.checkpatch is
missing").

Signed-off-by: Quentin Monnet <quentin@isovalent.com>