checkpatch: add new checkpatch image
Add a new Dockerfile to build an image containing the checkpatch.pl
script, along with a bash wrapper to configure and call it on the latest
commits or on the source files for Cilium's C code used in eBPF
programs.

Update the toplevel README.md, and add a dedicated README.md file in the
checkpatch directory.

Add custom patches that are applied to checkpatch.pl when building the
image, in order to fix false positives or to improve reports.

The bash script and the list of deprecated terms are copied from Cilium,
at commit 1843b4adedda ("build: run checkpatch.pl for bpf/ locally or as
GitHub action"). The bash script received minor modifications to add
copyright and licensing information, and to make it work in the Docker
container (it must not assume it is located in the Cilium repository and
change the work directory). Here is the complete diff:

    --- /path/to/cilium/contrib/checkpatch/checkpatch.sh
    +++ checkpatch.sh
    @@ -1,4 +1,6 @@
     #!/bin/bash
    +# SPDX-License-Identifier: Apache-2.0
    +# Copyright 2020 Authors of Cilium

     # Default options for checkpatch
     options="--no-tree --strict --no-summary --show-types --color=always"
    @@ -162,7 +164,6 @@
     echo "Retrieved $nb_commits commits $pr_info"
     echo

    -cd "$script_dir/../../"
     ret=0
     # Run checkpatch for BPF changes on all selected commits
     for ((i=0; i<$nb_commits; i++)); do

Note that this image does not relies on the other builder images in the
repository.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>