checkpatch: fix reported types for second run on files
The script checkpatch.pl behaves differently when run on Git commits or
on source code files. In the latter case, some reports are downgraded
from the warning level to simple checks. To account for those changes,
our shell script performs a second run to report those specifics issues
that have been downgraded.

Fix the list of issues to report:

- LONG_LINE_COMMENT should not be reported, we have it in the ingore
  list.
- Conversely, we want LONG_LINE_STRING, which we do not ignore.
- SPDX_LICENSE_TAG is only downgraded in a very specific case for the
  Linux kernel documentation, which does not apply here. Any occurrence
  of a missing tag will be reported on the first run already, do not
  report it on the second run as well.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>