checkpatch: ignore GIT_COMMIT_ID reports
Checkpatch complains when commits are not referenced the way it likes in
commit logs. While it is best to always follow the same formatting for
referencing commits, it is not always convenient to follow just one
form. We do not pay too much attention to this formatting, in part
because we work with GitHub and it adds links to commit hashes anyway.
Furthermore, reporting badly formatted commit references makes
checkpatch complain on backports touching bpf/ code (we always
reference the original commit, not in the form preferred by checkpatch).

Let's ignore such reports.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>