checkpatch: skip verification on backport PRs
Checkpatch should not be run on backport PRs. Style errors introduced in
the master branch should not be fixed during backports. We may miss a
few formatting errors introduced in conflict resolutions, but because
this is not the development branch anyway, this is a very minor
drawback.

Skip based on the name of the reference branch, when we run as a GitHub
action. 'trunk' or 'main' names are unused today, they're just added to
make the script more future-proof if the name of the main branch ever
changes.