images/maker: set GO111MODULE=off for go build commands
As of Go 1.16 the setting for GO111MODULE defaults to `on` which leads
to the `go build` commands in images/maker/build-go-deps.sh no longer
working. As a temporary measure set `GO111MODULE=off` for these commands
but add a TODO comment that they should use versioned `go install`
commands instead.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>