checkpatch: ignore FROM_SIGN_OFF_MISMATCH, report NO_AUTHOR_SIGN_OFF
We used to ignore NO_AUTHOR_SIGN_OFF reports because of false positives,
but the kernel checkpatch has changed for that check, and now it reports
when the sign-off tag is completely missing. Let's keep that check. The
reports for inconsistencies between author name and sign-off tag has
been moved to FROM_SIGN_OFF_MISMATCH, that we ignore.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>