checkpatch: run checkpatch on _all_ commits
Do not restrict checkpatch to commits that touch files under bpf/, but
run it on all commits. The objective is to run the tests on the subject
and log message for the commits. Whent commits do not touch files under
bpf/, just pass the log message to checkpatch. As a consequence, we need
to ignore the reports for NOT_UNIFIED_DIFF, which would be triggered
when processing empty diffs.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>