.github: only run workflows for pushes to master branch
Otherwise, when opening PRs from branches on
github.com/cilium/image-tools, each workflow gets instantiated twice,
once for the `on: pull_request` and once for `on: push`. Avoid this by
restricting `on: push` to pushes to the master branch.

FWIW, this is done in the same way on the github.com/cilium/cilium repo.

Signed-off-by: Tobias Klauser <tobias@cilium.io>