checkpatch: fix commit object parsing
The custom check on the length of the commit object has an issue, and
may fail to report long objects when the objects are on several lines,
as in the following example:

    ------
    example: This is the first line of the commit object,
    this is the second line of the commit object

    And here is the commit log message.

    Signed-off-by: John Doe <john.doe@domain.tld>
    ------

The failure does not occur on local runs of the script, it only happens
when running as a GitHub action. This is because the commit object is
retrieved differently in the two cases:

- When running locally, we git it from `git log --pretty=format:...`, so
  in other words, we take the object as considered by git, which is the
  correct thing to do.

- When running as a GitHub action, we retrieve a single multi-line block
  from the GitHub API, containing both the object and the log message
  for the commit. It falls down on us to parse this block to extract the
  commit object. So far we would only consider the first line of the
  block, so from the example above, we would extract the first line
  ending at the comma.

  But git wants a blank line between the object and the log message, so
  this parsing is, in fact, incorrect. We need to add a second blank
  line to the delimiter, in order to retrieve the full object for the
  commit. This is what we do here.

After this change, the GitHub action should correctly report when commit
objects are too long (sometimes as a result of the author omitting the
blank line between the object and the log message).

Signed-off-by: Quentin Monnet <quentin@isovalent.com>