checkpatch: fix return status for GitHub action
When running checkpatch as a GitHub action, we pipe a diff into
checkpatch.pl, which return a non-null exit status if issues are
reported.

We want to propagate this exit status and exit from the script with it,
so that the GitHub action marks the test as failed in such a case. This
is what we did initially. However, a commit piped the output of
checkpatch.pl into a sed filter. This had the unforeseen effect of
replacing the return code of the whole command pipeline with the one
from the sed filter (instead of the one from checkpatch.pl). As a
consequence, checkpatch.pl exit status is ignored in favour of sed's,
and reporting issues no longer results in a failure of the GitHub
action.

Let's fix by making sure the exit status from checkpatch.pl, and not
from sed, is returned from the subshell to be propagated.

Fixes: 30ff1420a667 ("checkpatch: format errors/warnings and group log lines for GH Action")
Signed-off-by: Quentin Monnet <quentin@isovalent.com>