checkpatch: Ignore BIT_MACRO check reports
Checkpatch CI workflow has failed in the past with:

    BIT_MACRO: Prefer using the BIT macro

Although it is recommended to use this macro in kernel code, we do not
use it (or have it defined) in Cilium's BPF code. This check is not
relevant for us, so let's disable it.

Copyright years in header are not up-to-date, remove years altogether.

Suggested-by: Maxim Mikityanskiy <maxim@isovalent.com>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>