checkpatch: Add root .gitconfig
Add root .gitconfig which sets all directories as safe. This prevents git
failure when running on volumes mounted on macOS, for example (which show
root as the owner of the mounted directory within the container, see
docker/for-mac#2657).

Since checkpatch image is intended to be used with a Cilium repo mounted
on a root directory (e.g., '/workspace') there is no risk of other users
having untrusted '.git/config' files between '/.gitconfig' (added in this
commit), and '/workspace/.git/config' (mounted from the Cilium
repository), so there is no risk running malicious commands as speculated
in
https://github.blog/2022-04-12-git-security-vulnerability-announced/#cve-2022-24765.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>