Fix maker-image's circular build dependency
This commit adjusts `scripts/build-image.sh` to avoid attempting to run
the maker-image while trying to build the maker-image. If a user is
attempting to build the maker-image and it is detected that the build
is running outside of the maker-image, then an attempt to run the
maker-image will not be made. Users will have to ensure that local
versions of the required tools are available.

The build-image.sh script will attempt to run crane within the
maker-image, in order to pull information about the target build image.
If the user is trying to build the maker-image, and a pre-existing
maker-image is not available, for instance if quay.io is down, then
the script will be unable to build a new maker-image. In this case,
the user is not expected to have a local copy of crane that the script
can use.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>