checkpatch: Remove grouping logs in GHA output
Grouping logs by commit in the output of the GitHub workflow provides a
nicer interface than just displaying separators, but it turns out to be
more cumbersome in case of errors, because users have no way to know
what group contains the error and they have to manually expand all
groups, which has proved annoying. There doesn't seem to be an option to
auto-expand the groups, so let's remove them altogether.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>