ci: build images on pull requests
Currently images are only built (and pushed) on pushes to the master
branch, i.e. once the PR is merged. This can lead to situations where
the image build is broken by a PR but the breakage is only caught later
on, see e.g. [1] and [2].

Fix this by also building the images on PRs but only push them to the
registry for master pushes as previously.

[1] https://github.com/cilium/image-tools/pull/233#discussion_r1532577369
[2] https://github.com/cilium/image-tools/pull/240

Signed-off-by: Tobias Klauser <tobias@cilium.io>