images: Add clang-format image
This commit adds a clang-format image that can be used to format C code.
The image is made to be as minimal as possible, just a statically
compiled clang-format binary in a scratch container. This makes it
quick to pull and allows for the extraction of the binary to run on
a host system.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>