renovate: Group base images updates
Currently a couple of docker images use `latest` or an equivalently
mutable tag that gets republished almost daily, this leads to renovate
constantly trying to update the digest for those images. In practice it
means that as soon as we merge a renovate PR for a given image, renovate
creates a new one for that same image.

With this PR we introduce two changes to the renovate config concerning
docker image updates:
* Group them all together, meaning there will be a single PR for all
  base images updates instead of one per base image
* Only do these updates monthly (I'm open to any other suggested
  frequency, the main goal is to avoid having a new PR created as soon
  as one is merged, although #335 should help with the PR toil)

Signed-off-by: Hadrien Patte <hadrien.patte@datadoghq.com>