checkpatch: re-enable verification of PRs against all branches
1392d2f42d82 ("checkpatch: skip verification on backport PRs") disabled
running checkpatch against stable branches, as at that time the check
turned out being too noisy on backport PRs, which is also not the right
place to fix style errors.

However, it's been a long time since then, and the expectation is that
both the main and stable branches are in a much better shape now, which
should strongly reduce the possibility of hitting failures during backports.
Hence, I'd argue it is time to enable checkpatch for all branches again,
which has the nice side-effects of reducing the likelihood of silent
breakages if the default branch does not match one of the previously
hardcoded ones, as well as getting early signals when working e.g.,
against feature branches, rather than having to wait for feedback until
actually targeting the main branch.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>