checkpatch: reduce likelihood of issues due to GH pagination
By default, the commits API returns 30 results per-page. Let's increase
this to the maximum possible, that is 100, to reduce the likelihood of
actually hitting this limit, without having to implement proper handling
of pagination. While I've seen PRs with 30ish commits in the past, I
don't think I've ever seen one with 100 commits, so I think we should
be fine from that point of view. I would expect reviewers to complain
anyhow with that many commits.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>