checkpatch: prevent unbound variable error
The blamed commit set the option to make the script fail if it tries to
access any unbound variables, to catch possible errors. However, the
script now fails when invoked directly, due to:

> ./checkpatch.sh: line 218: GITHUB_REF: unbound variable

While this is not a problem in practice, given that the Cilium's
Makefile always sets the GITHUB_REF variable (possibly empty), let's
guard the variable accesses, to be a bit more user-friendly.

Fixes: 5d11067d6e07 ("checkpatch: make script exit if command exits with error")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>